first working run

+++ src/range.h
@@ -1,10 +1,12 @@
+#include <stddef.h>
+
 #ifndef RANGE_H
 #define RANGE_H
 
 struct range
 {
-  unsigned int start;
-  unsigned int end;
+  size_t start;
+  size_t end;
 };
 
 #endif // RANGE_H
